[#116] Ensure that unique names are used
When updating a dataset, always use the existing name, got it from the
db via the dataset id. If not found abort. When creating a dataset, get
it from the task, and if not there (eg. initial run of the changelog
harvester, which we shouldn't do anyway) slugify the title.